feat(launchpad): move terminal detection into tui
