+++ sn_node_launchpad/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 authors = ["MaidSafe Developers <dev@maidsafe.net>"]
 description = "Node Launchpad"
-name = "sn_node_launchpad"
+name = "node_launchpad"
 version = "0.1.0"
 edition = "2021"
 license = "GPL-3.0"
@@ -14,9 +14,6 @@ build = "build.rs"
 name = "node-launchpad"
 path = "src/bin/tui/main.rs"
 
-[[bin]]
-name = "autonomi-launcher"
-path = "src/bin/launcher/main.rs"
 
 [dependencies]
 better-panic = "0.3.0"
